feat: run in WebAssembly (Pyodide) without numba
numba has no WebAssembly build, and as a hard dependency it made
fastwoe uninstallable in Pyodide. It is now installed everywhere except
there (numba; sys_platform != "emscripten").

Without numba, the Somers' D functions run as written, as plain Python:
the calculation is not replaced, and the integer pair counts are
identical to the compiled version (tested by running the same cases
with numba and with it blocked, and in Pyodide against native numba).

In Pyodide 0.27.7: Somers' D on 200,000 rows in 1.1s (binary target)
to 3.3s (continuous); FastWoe.fit on 10 numeric features x 50,000 rows
in 1.6s (0.7s natively with numba). No warning is shown under
WebAssembly, where the missing numba is expected.

Also fixes a 64-bit integer assumption in the conditional transform
(np.bincount on int64 group codes) that failed on 32-bit wasm.

+++ fastwoe/fastwoe_conditional.py
@@ -323,7 +323,7 @@ class ConditionalWoeMixin:
         Xb = self._binned_frame(X[order])
 
         groups = Xb.groupby(order, dropna=False, sort=False, observed=True)
-        codes = groups.ngroup().to_numpy()
+        codes = groups.ngroup().to_numpy(dtype=np.intp)  # native int: wasm32 is 32-bit
         sizes = np.bincount(codes)
         firsts = groups.head(1)
